add default and examples to config json schema
this makes most editors autocomplete the default value when you want to override,
acting as an in-ide documentation for the keybindings instead of forcing people to
go to the website.

the only thing I dislike about it is that it is repeating stuff from tui,
and the proof for the fragility of this solution is that the 'help' keybinding
didn't work well.

maybe we can use ts->go generation, or go->ts codegen to ensure
they are always in sync.

but that's an issue for a different PR I guess.