fix: stop false positives for @file references (#2198)
This is a @valid/path/to/a/file, and it should also match at the beginning of a
line:

@another-valid/path/to/a/file

but references to email addresses and other strings are not:

   - Adds a "Co-authored-by:" footer which clarifies which AI agent
     helped create this commit, using an appropriate `noreply@...`
     email address.

However currently the fileRegex is `/@([^\s]+)/g` which is much too inclusive
and triggers false positives.

As a first approximation, we can use a negative look-behind assertion to filter
out any matches for which the `@` is immediately preceded by an alphanumeric
character, and then also impose some heuristics regarding use of backticks,
commands, periods, and whitespace.

This can be seen as functional at: https://regex101.com/r/F6KAgH/5

Fixes #2198.