docs: fix gopls LSP name in config
Using `gopls` in the `opencode.json` config isn't working, instead we
need to use `golang`.

The matching is currently based on the `id` field in the LSP servers:
https://github.com/rekram1-node/opencode-sst/blob/73792b37b122a87cae837178969ee82c61a904cb/packages/opencode/src/lsp/server.ts#L190

Although the column title seems to refer to the underlying LSP server name (e.g.
for pyright), this naming follows the way `rust` LSP is documented below.