fix: Only block paths outside working dir if non prompting
This modifies the blocking of access outside of the working dir to allow `ask`
permissions since the user has an opportunity to review the command being
executed.

I often go outside of the working dir if I have the model create files in /tmp
or am debugging a system issue where it needs to read config files in /etc.