Fix broken version parsing in installation script
The installation script fails for me with the following error message:

```
> curl -fsSL https://opencode.ai/install | bash
Failed to fetch version information
```

The culprit is the reliance on a space in the targeted `"tag_name": "`
string segment when going through the releases API response.

Currently, the API response contains this version of the string:
```
"tag_name":"v1.11.0"
```

As such, the regular expression with the hardcoded whitespace fails to
match this.

This PR adds an optional whitespace marker into the regular expression
so that it will match the `tag_name` both with or without a space
in-between key and value.