Switch to a more portable sed pattern matching
The `awk` pattern matching had several issues:

- the original version assumed specific whitespace, which could not be
  relied upon with GitHub APIs
- when adapting the pattern to handle whitespace, the target group
  changed positioning, so `$4` was relative depending on whitespace
- whitespace handling in awk differs between POSIX and MacOS X.

This PR changes the pattern matching to a `sed`-based format which
should be more portable. It should work across popular POSIX distros as
well as MacOS X, and it should be able to handle abitrary variations on
the whitespace between key and value.