fix(publish): Fetch tags from origin before creating GitHub release
A few publishes have failed in cases where "tag exists locally but not on remote", if another fetch is performed it should update local->remote tag references. The push should handle this already but it can't hurt to have it explicitly update right?

I also thought that it could be a timing issue or a problem with the `gh` command itself (if it uses some other method to check remote tags).

If I'm completely wrong, apologies for wasting your time.