fix(theme): filter out null values from theme palette
**Issue**

When opening opencode in zellij we see the following error:

```
null is not an object (evaluating 'hex3.replace')
```

This is because opencode requests 16 colors from the terminal but only
10 are returned when requesting those colors from zellij.

This doesn't seem to break much in opencode, but the big red error log
is quite annoying.

**Solution**

Because zellij usually returns 10 colors and looking at the
implementation here in opencode only 8 colors are required, we can
safely filter out the null values.

Fixes #4017

Signed-off-by: Kamaal Farah <kamaal.f1@gmail.com>