feat: add max steps for supervisor and sub-agents
We want a way to limit the number of steps allow
agents in the app can take. This is primarily useful
for sub-agents, but it applies to the supervisor as
well.

The user can define a maxSteps value for each agent.

Documentation has also been updated for this new feature.